+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+# Unreleased
+
+- certwrapper will now exit if it can't write to any of the required paths before attempting to
+  obtain a certificate.
+- Update to Go 1.17
+
 # v2.0.0
 
 - Configuration of certwrapper can now be done using environment variables.